Separate spell-related classes
