+++ game/world/managers/objects/spell/SpellManager.py
@@ -1,385 +1,18 @@
 import time
-import random
 from struct import pack
 
 from database.dbc.DbcDatabaseManager import DbcDatabaseManager
-from database.dbc.DbcModels import Spell, SpellCastTimes, SpellRange, SpellDuration
 from database.realm.RealmDatabaseManager import RealmDatabaseManager, CharacterSpell
 from database.world.WorldDatabaseManager import WorldDatabaseManager
-from game.world.managers.abstractions.Vector import Vector
 from game.world.managers.maps.MapManager import MapManager
-from game.world.managers.objects.ObjectManager import ObjectManager
-from game.world.managers.objects.player.DuelManager import DuelManager
-from game.world.managers.objects.spell.EffectTargets import EffectTargets
+from game.world.managers.objects.spell.CastingSpell import CastingSpell
+from game.world.managers.objects.spell.SpellEffectHandler import SpellEffectHandler
 from network.packet.PacketWriter import PacketWriter, OpCode
-from utils.Logger import Logger
 from utils.constants.ItemCodes import InventoryError, InventoryTypes
-from utils.constants.ObjectCodes import ObjectTypes, AttackTypes
+from utils.constants.ObjectCodes import ObjectTypes
 from utils.constants.SpellCodes import SpellCheckCastResult, SpellCastStatus, \
-    SpellMissReason, SpellTargetMask, SpellState, SpellEffects, SpellTargetType, SpellAttributes, SpellAttributesEx, \
-    AuraTypes, SpellCastFlags
-from utils.constants.UnitCodes import PowerTypes, UnitFlags
-
-
-class CastingSpell(object):
-    spell_entry: Spell
-    cast_state: SpellState
-    cast_flags: SpellCastFlags  # TODO Write proc flag when needed
-    spell_caster = None  # TODO Item caster (use item?)
-    initial_target = None
-    unit_target_results: dict  # Assigned on cast - contains guids and results on successful hits/misses/blocks etc.
-    spell_target_mask: SpellTargetMask
-    range_entry: SpellRange
-    duration_entry: SpellDuration
-    cast_time_entry: SpellCastTimes
-    effects: list
-
-    cast_end_timestamp: float
-    spell_delay_end_timestamp: float
-    caster_effective_level: int
-
-    spell_attack_type: int
-
-    def __init__(self, spell, caster_obj, initial_target, target_mask):
-        self.spell_entry = spell
-        self.spell_caster = caster_obj
-        self.initial_target = initial_target
-        self.spell_target_mask = target_mask
-        self.duration_entry = DbcDatabaseManager.spell_duration_get_by_id(spell.DurationIndex)
-        self.range_entry = DbcDatabaseManager.spell_range_get_by_id(spell.RangeIndex)
-        self.cast_time_entry = DbcDatabaseManager.spell_cast_time_get_by_id(spell.CastingTimeIndex)
-        self.cast_end_timestamp = self.get_base_cast_time()/1000 + time.time()
-        self.caster_effective_level = self.calculate_effective_level(self.spell_caster.level)
-
-        self.spell_attack_type = AttackTypes.RANGED_ATTACK if self.is_ranged() else AttackTypes.BASE_ATTACK
-        self.cast_state = SpellState.SPELL_STATE_PREPARING
-        self.effects = self.load_effects()
-
-        self.cast_flags = SpellCastFlags.CAST_FLAG_NONE  # TODO Ammo/proc flag
-
-    def initial_target_is_object(self):
-        return isinstance(self.initial_target, ObjectManager)
-
-    def initial_target_is_unit_or_player(self):
-        if not self.initial_target_is_object():
-            return False
-
-        target_type = self.initial_target.get_type()
-        return target_type == ObjectTypes.TYPE_UNIT or target_type == ObjectTypes.TYPE_PLAYER
-
-    def initial_target_is_player(self):
-        if not self.initial_target_is_object():
-            return False
-
-        return self.initial_target.get_type() == ObjectTypes.TYPE_PLAYER
-
-    def initial_target_is_item(self):
-        if not self.initial_target_is_object():
-            return False
-
-        return self.initial_target.get_type() == ObjectTypes.TYPE_ITEM
-
-    def initial_target_is_gameobject(self):
-        if not self.initial_target_is_object():
-            return False
-
-        return self.initial_target.get_type() == ObjectTypes.TYPE_GAMEOBJECT
-
-    def initial_target_is_terrain(self):
-        return isinstance(self.initial_target, Vector)
-
-    def get_target_info(self):  # ([values], len)
-        is_terrain = self.initial_target_is_terrain()
-        return ([self.initial_target.x, self.initial_target.y, self.initial_target.z] if is_terrain
-                else [self.initial_target.guid]), ('3f' if is_terrain else 'Q')
-
-    def is_instant_cast(self):
-        return self.cast_time_entry.Base == 0
-
-    def is_ranged(self):
-        return self.spell_entry.Attributes & SpellAttributes.SPELL_ATTR_RANGED == SpellAttributes.SPELL_ATTR_RANGED
-
-    def is_passive(self):
-        return self.spell_entry.Attributes & SpellAttributes.SPELL_ATTR_PASSIVE == SpellAttributes.SPELL_ATTR_PASSIVE
-
-    def trigger_cooldown_on_remove(self):
-        return self.spell_entry.Attributes & SpellAttributes.SPELL_ATTR_DISABLED_WHILE_ACTIVE == SpellAttributes.SPELL_ATTR_DISABLED_WHILE_ACTIVE
-
-    def casts_on_swing(self):
-        return self.spell_entry.Attributes & SpellAttributes.SPELL_ATTR_ON_NEXT_SWING_1 == SpellAttributes.SPELL_ATTR_ON_NEXT_SWING_1
-
-    def requires_combo_points(self):
-        cp_att = SpellAttributesEx.SPELL_ATTR_EX_REQ_TARGET_COMBO_POINTS | SpellAttributesEx.SPELL_ATTR_EX_REQ_COMBO_POINTS
-        return self.spell_caster.get_type() == ObjectTypes.TYPE_PLAYER and \
-            self.spell_entry.AttributesEx & cp_att != 0
-
-    def calculate_effective_level(self, level):
-        if level > self.spell_entry.MaxLevel > 0:
-            level = self.spell_entry.MaxLevel
-        elif level < self.spell_entry.BaseLevel:
-            level = self.spell_entry.BaseLevel
-        return level - self.spell_entry.SpellLevel
-
-    def get_base_cast_time(self):
-        skill = self.spell_caster.skill_manager.get_skill_for_spell_id(self.spell_entry.ID)
-        if not skill:
-            return self.cast_time_entry.Minimum
-
-        return int(max(self.cast_time_entry.Minimum, self.cast_time_entry.Base + self.cast_time_entry.PerLevel * skill.value))
-
-    def get_resource_cost(self):
-        if self.spell_caster.get_type() == ObjectTypes.TYPE_PLAYER and self.spell_entry.ManaCostPct != 0:
-            return self.spell_caster.base_mana * self.spell_entry.ManaCostPct / 100
-
-        # ManaCostPerLevel is not used by anything relevant (only 271/4513/7290)
-        return self.spell_entry.ManaCost
-
-    def load_effects(self):
-        effects = []
-        if self.spell_entry.Effect_1 != 0:
-            effects.append(SpellEffect(self, 1))
-        if self.spell_entry.Effect_2 != 0:
-            effects.append(SpellEffect(self, 2))
-        if self.spell_entry.Effect_3 != 0:
-            effects.append(SpellEffect(self, 3))
-        return effects
-    
-    def get_reagents(self):
-        return (self.spell_entry.Reagent_1, self.spell_entry.ReagentCount_1), (self.spell_entry.Reagent_2, self.spell_entry.ReagentCount_2), \
-               (self.spell_entry.Reagent_3, self.spell_entry.ReagentCount_3), (self.spell_entry.Reagent_4, self.spell_entry.ReagentCount_4), \
-               (self.spell_entry.Reagent_5, self.spell_entry.ReagentCount_5), (self.spell_entry.Reagent_6, self.spell_entry.ReagentCount_6), \
-               (self.spell_entry.Reagent_7, self.spell_entry.ReagentCount_7), (self.spell_entry.Reagent_8, self.spell_entry.ReagentCount_8)
-
-    def get_conjured_items(self):
-        conjured_items = []
-        for effect in self.effects:
-            item_count = abs(effect.get_effect_points(self.caster_effective_level))
-            conjured_items.append([effect.item_type, item_count])
-        return tuple(conjured_items)
-
-
-class SpellEffect(object):
-    effect_type: SpellEffects
-    die_sides: int
-    base_dice: int
-    dice_per_level: int
-    real_points_per_level: int
-    base_points: int
-    implicit_target_a: int
-    implicit_target_b: int
-    radius_index: int
-    aura_type: int
-    aura_period: int
-    amplitude: int
-    chain_targets: int
-    item_type: int
-    misc_value: int
-    trigger_spell: int
-
-    effect_index: int
-    targets: EffectTargets
-
-    def __init__(self, casting_spell, index):
-        if index == 1:
-            self.load_first(casting_spell.spell_entry)
-        elif index == 2:
-            self.load_second(casting_spell.spell_entry)
-        elif index == 3:
-            self.load_third(casting_spell.spell_entry)
-
-        self.targets = EffectTargets(casting_spell, self)
-
-    def get_effect_points(self, effective_level):
-        rolled_points = random.randint(1, self.die_sides + self.dice_per_level) if self.die_sides != 0 else 0
-        return self.base_points + int(self.real_points_per_level * effective_level) + rolled_points
-
-    def load_first(self, spell):
-        self.effect_type = spell.Effect_1
-        self.die_sides = spell.EffectDieSides_1
-        self.base_dice = spell.EffectBaseDice_1
-        self.dice_per_level = spell.EffectDicePerLevel_1
-        self.real_points_per_level = spell.EffectRealPointsPerLevel_1
-        self.base_points = spell.EffectBasePoints_1
-        self.implicit_target_a = spell.ImplicitTargetA_1
-        self.implicit_target_b = spell.ImplicitTargetB_1
-        self.radius_index = spell.EffectRadiusIndex_1
-        self.aura_type = spell.EffectAura_1
-        self.aura_period = spell.EffectAuraPeriod_1
-        self.amplitude = spell.EffectAmplitude_1
-        self.chain_targets = spell.EffectChainTargets_1
-        self.item_type = spell.EffectItemType_1
-        self.misc_value = spell.EffectMiscValue_1
-        self.trigger_spell = spell.EffectTriggerSpell_1
-
-        self.effect_index = 1
-        
-    def load_second(self, spell):
-        self.effect_type = spell.Effect_2
-        self.die_sides = spell.EffectDieSides_2
-        self.base_dice = spell.EffectBaseDice_2
-        self.dice_per_level = spell.EffectDicePerLevel_2
-        self.real_points_per_level = spell.EffectRealPointsPerLevel_2
-        self.base_points = spell.EffectBasePoints_2
-        self.implicit_target_a = spell.ImplicitTargetA_2
-        self.implicit_target_b = spell.ImplicitTargetB_2
-        self.radius_index = spell.EffectRadiusIndex_2
-        self.aura_type = spell.EffectAura_2
-        self.aura_period = spell.EffectAuraPeriod_2
-        self.amplitude = spell.EffectAmplitude_2
-        self.chain_targets = spell.EffectChainTargets_2
-        self.item_type = spell.EffectItemType_2
-        self.misc_value = spell.EffectMiscValue_2
-        self.trigger_spell = spell.EffectTriggerSpell_2
-
-        self.effect_index = 2
-
-    def load_third(self, spell):
-        self.effect_type = spell.Effect_3
-        self.die_sides = spell.EffectDieSides_3
-        self.base_dice = spell.EffectBaseDice_3
-        self.dice_per_level = spell.EffectDicePerLevel_3
-        self.real_points_per_level = spell.EffectRealPointsPerLevel_3
-        self.base_points = spell.EffectBasePoints_3
-        self.implicit_target_a = spell.ImplicitTargetA_3
-        self.implicit_target_b = spell.ImplicitTargetB_3
-        self.radius_index = spell.EffectRadiusIndex_3
-        self.aura_type = spell.EffectAura_3
-        self.aura_period = spell.EffectAuraPeriod_3
-        self.amplitude = spell.EffectAmplitude_3
-        self.chain_targets = spell.EffectChainTargets_3
-        self.item_type = spell.EffectItemType_3
-        self.misc_value = spell.EffectMiscValue_3
-        self.trigger_spell = spell.EffectTriggerSpell_3
-
-        self.effect_index = 3
-    
-
-class SpellEffectHandler(object):
-    @staticmethod
-    def apply_effect(casting_spell, effect, target):
-        if effect.effect_type not in SPELL_EFFECTS:
-            Logger.debug(f'Unimplemented effect called: {effect.effect_type}')
-            return
-        SPELL_EFFECTS[effect.effect_type](casting_spell, effect, casting_spell.spell_caster, target)
-
-    @staticmethod
-    def handle_school_damage(casting_spell, effect, caster, target):
-        damage = effect.get_effect_points(casting_spell.caster_effective_level)
-        caster.deal_spell_damage(target, damage, casting_spell.spell_entry.School, casting_spell.spell_entry.ID)
-
-    @staticmethod
-    def handle_heal(casting_spell, effect, caster, target):
-        healing = effect.get_effect_points(casting_spell.caster_effective_level)
-
-    @staticmethod
-    def handle_weapon_damage(casting_spell, effect, caster, target):
-        damage_info = caster.calculate_melee_damage(target, casting_spell.spell_attack_type)
-        if not damage_info:
-            return
-        damage = damage_info.total_damage + effect.get_effect_points(casting_spell.caster_effective_level)
-        caster.deal_spell_damage(target, damage, casting_spell.spell_entry.School, casting_spell.spell_entry.ID)
-
-    @staticmethod
-    def handle_weapon_damage_plus(casting_spell, effect, caster, target):
-        damage_info = caster.calculate_melee_damage(target, casting_spell.spell_attack_type)
-        if not damage_info:
-            return
-        damage = damage_info.total_damage
-        damage_bonus = effect.get_effect_points(casting_spell.caster_effective_level)
-
-        if caster.get_type() == ObjectTypes.TYPE_PLAYER and \
-                casting_spell.requires_combo_points():
-            damage_bonus *= caster.combo_points
-
-        caster.deal_spell_damage(target, damage + damage_bonus, casting_spell.spell_entry.School, casting_spell.spell_entry.ID)
-
-    @staticmethod
-    def handle_add_combo_points(casting_spell, effect, caster, target):
-        caster.add_combo_points_on_target(target, effect.get_effect_points(casting_spell.caster_effective_level))
-
-    @staticmethod
-    def handle_aura_application(casting_spell, effect, caster, target):
-        target.aura_manager.apply_spell_effect_aura(caster, casting_spell, effect)
-
-    @staticmethod
-    def handle_request_duel(casting_spell, effect, caster, target):
-        duel_result = DuelManager.request_duel(caster, target, effect.misc_value)
-        if duel_result == 1:
-            result = SpellCheckCastResult.SPELL_NO_ERROR
-        elif duel_result == 0:
-            result = SpellCheckCastResult.SPELL_FAILED_TARGET_DUELING
-        else:
-            result = SpellCheckCastResult.SPELL_FAILED_DONT_REPORT
-        caster.spell_manager.send_cast_result(casting_spell.spell_entry.ID, result)
-
-    @staticmethod
-    def handle_energize(casting_spell, effect, caster, target):
-        power_type = effect.misc_value
-
-        if power_type != target.power_type:
-            return
-
-        new_power = target.get_power_type_value() + effect.get_effect_points(casting_spell.caster_effective_level)
-        if power_type == PowerTypes.TYPE_MANA:
-            target.set_mana(new_power)
-        elif power_type == PowerTypes.TYPE_RAGE:
-            target.set_rage(new_power)
-        elif power_type == PowerTypes.TYPE_FOCUS:
-            target.set_focus(new_power)
-        elif power_type == PowerTypes.TYPE_ENERGY:
-            target.set_energy(new_power)
-
-    @staticmethod
-    def handle_summon_mount(casting_spell, effect, caster, target):
-        already_mounted = target.unit_flags & UnitFlags.UNIT_MASK_MOUNTED
-        if already_mounted:
-            # Remove any existing mount auras.
-            target.aura_manager.remove_auras_by_type(AuraTypes.SPELL_AURA_MOUNTED)
-            target.aura_manager.remove_auras_by_type(AuraTypes.SPELL_AURA_MOD_INCREASE_MOUNTED_SPEED)
-            # Force dismount if target is still mounted (like a previous SPELL_EFFECT_SUMMON_MOUNT that doesn't
-            # leave any applied aura).
-            if target.mount_display_id > 0:
-                target.unmount()
-                target.set_dirty()
-        else:
-            creature_entry = effect.misc_value
-            if not target.summon_mount(creature_entry):
-                Logger.error(f'SPELL_EFFECT_SUMMON_MOUNT: Creature template ({creature_entry}) not found in database.')
-
-    @staticmethod
-    def handle_insta_kill(casting_spell, effect, caster, target):
-        # No SMSG_SPELLINSTAKILLLOG in 0.5.3
-        target.die(killer=caster)
-
-    @staticmethod
-    def handle_create_item(casting_spell, effect, caster, target):
-        if target.get_type() != ObjectTypes.TYPE_PLAYER:
-            return
-
-        target.inventory.add_item(effect.item_type,
-                                  count=effect.get_effect_points(casting_spell.caster_effective_level))
-
-    @staticmethod
-    def handle_teleport_units(casting_spell, effect, caster, target):
-        teleport_info = effect.targets.implicit_target_b
-        target.teleport(teleport_info[0], teleport_info[1])  # map, coordinates resolved
-        # TODO Die sides are assigned for at least Word of Recall (ID 1)
-
-
-SPELL_EFFECTS = {
-    SpellEffects.SPELL_EFFECT_SCHOOL_DAMAGE: SpellEffectHandler.handle_school_damage,
-    SpellEffects.SPELL_EFFECT_HEAL: SpellEffectHandler.handle_heal,
-    SpellEffects.SPELL_EFFECT_WEAPON_DAMAGE: SpellEffectHandler.handle_weapon_damage,
-    SpellEffects.SPELL_EFFECT_ADD_COMBO_POINTS: SpellEffectHandler.handle_add_combo_points,
-    SpellEffects.SPELL_EFFECT_DUEL: SpellEffectHandler.handle_request_duel,
-    SpellEffects.SPELL_EFFECT_WEAPON_DAMAGE_PLUS: SpellEffectHandler.handle_weapon_damage_plus,
-    SpellEffects.SPELL_EFFECT_APPLY_AURA: SpellEffectHandler.handle_aura_application,
-    SpellEffects.SPELL_EFFECT_ENERGIZE: SpellEffectHandler.handle_energize,
-    SpellEffects.SPELL_EFFECT_SUMMON_MOUNT: SpellEffectHandler.handle_summon_mount,
-    SpellEffects.SPELL_EFFECT_INSTAKILL: SpellEffectHandler.handle_insta_kill,
-    SpellEffects.SPELL_EFFECT_CREATE_ITEM: SpellEffectHandler.handle_create_item,
-    SpellEffects.SPELL_EFFECT_TELEPORT_UNITS: SpellEffectHandler.handle_teleport_units
-}
+    SpellMissReason, SpellTargetMask, SpellState, SpellEffects, SpellAttributes, SpellCastFlags
+from utils.constants.UnitCodes import PowerTypes
 
 
 class SpellManager(object):
